backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jinja2
python-multipart
aiofiles
pydantic[email]==2.9.2
sqlalchemy
psycopg2-binary
passlib[bcrypt]==1.7.4
python-jose==3.3.0
alembic==1.12.0
python-dotenv==1.1.0
pydantic-settings
PyJWT>=2.8.0
httpx>=0.24.0



